# openapi/api-server.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/auth/github/start": {
      "get": {
        "operationId": "AuthController_startGitHub",
        "parameters": [
          {
            "name": "redirectTo",
            "required": false,
            "in": "query",
            "schema": {
              "example": "/posts",
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "GitHub OAuth 승인 화면으로 리다이렉트한다."
          }
        },
        "summary": "GitHub OAuth 로그인 시작",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/github/callback": {
      "get": {
        "operationId": "AuthController_githubCallback",
        "parameters": [
          {
            "name": "error",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "code",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "세션 쿠키를 설정한 뒤 web 화면으로 리다이렉트한다."
          }
        },
        "summary": "GitHub OAuth callback 처리",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/signup/complete": {
      "post": {
        "operationId": "AuthController_completeSignUp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteSignUpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "sessionCookie": []
          }
        ],
        "summary": "OAuth 가입 완료",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "sessionCookie": []
          }
        ],
        "summary": "현재 로그인 사용자 조회",
        "tags": [
          "auth"
        ]
      },
      "patch": {
        "operationId": "AuthController_updateMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCurrentUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "sessionCookie": []
          }
        ],
        "summary": "현재 사용자 정보 수정",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "sessionCookie": []
          }
        ],
        "summary": "로그아웃",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/posts/{postId}/comments": {
      "get": {
        "operationId": "CommentsController_findComments",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "post-1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentListResponseDto"
                }
              }
            }
          }
        },
        "summary": "댓글 목록 조회",
        "tags": [
          "comments"
        ]
      },
      "post": {
        "operationId": "CommentsController_createComment",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "post-1",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "sessionCookie": []
          }
        ],
        "summary": "댓글 작성",
        "tags": [
          "comments"
        ]
      }
    },
    "/api/posts/{postId}/comments/{commentId}": {
      "patch": {
        "operationId": "CommentsController_updateComment",
        "parameters": [
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "comment-1",
              "type": "string"
            }
          },
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "post-1",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "sessionCookie": []
          }
        ],
        "summary": "댓글 수정",
        "tags": [
          "comments"
        ]
      },
      "delete": {
        "operationId": "CommentsController_deleteComment",
        "parameters": [
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "comment-1",
              "type": "string"
            }
          },
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "post-1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteCommentResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "sessionCookie": []
          }
        ],
        "summary": "댓글 삭제",
        "tags": [
          "comments"
        ]
      }
    },
    "/api/post-tags": {
      "get": {
        "description": "태그는 별도 PostTag 데이터로 관리한다. 일반 사용자용 태그 생성 API는 제공하지 않는다.",
        "operationId": "PostTagsController_findTags",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PostTagDto"
                  }
                }
              }
            }
          }
        },
        "summary": "게시글 태그 목록 조회",
        "tags": [
          "post-tags"
        ]
      }
    },
    "/api/posts": {
      "get": {
        "description": "검색어, 태그, 페이지, 정렬, 보기 옵션을 받아 게시글 목록을 반환한다.",
        "operationId": "PostsController_findPosts",
        "parameters": [
          {
            "name": "view",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "table",
                "card"
              ],
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "created-desc",
                "created-asc",
                "title-asc"
              ],
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "tagId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostListResponseDto"
                }
              }
            }
          }
        },
        "summary": "게시글 목록 조회",
        "tags": [
          "posts"
        ]
      },
      "post": {
        "operationId": "PostsController_createPost",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "sessionCookie": []
          }
        ],
        "summary": "게시글 생성",
        "tags": [
          "posts"
        ]
      }
    },
    "/api/posts/{id}": {
      "get": {
        "operationId": "PostsController_findPost",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "post-1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostDto"
                }
              }
            }
          }
        },
        "summary": "게시글 단건 조회",
        "tags": [
          "posts"
        ]
      },
      "patch": {
        "operationId": "PostsController_updatePost",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "post-1",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "sessionCookie": []
          }
        ],
        "summary": "게시글 수정",
        "tags": [
          "posts"
        ]
      },
      "delete": {
        "operationId": "PostsController_deletePost",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "post-1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletePostResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "sessionCookie": []
          }
        ],
        "summary": "게시글 삭제",
        "tags": [
          "posts"
        ]
      }
    },
    "/api/health": {
      "get": {
        "operationId": "HealthController_health",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "health"
        ]
      }
    }
  },
  "info": {
    "title": "NMM API",
    "description": "나만의 무기 만들기 보일러플레이트용 게시판 API",
    "version": "0.1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "nmm_session"
      },
      "session": {
        "scheme": "bearer",
        "bearerFormat": "session",
        "type": "http"
      }
    },
    "schemas": {
      "CompleteSignUpDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "sijun"
          }
        },
        "required": [
          "name"
        ]
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "user-1"
          },
          "email": {
            "type": "string",
            "example": "sijun@example.com"
          },
          "name": {
            "type": "string",
            "nullable": true,
            "example": "sijun"
          },
          "image": {
            "type": "string",
            "nullable": true,
            "example": "https://avatars.githubusercontent.com/u/1?v=4"
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "ADMIN"
            ],
            "example": "USER"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "ACTIVE",
              "SUSPENDED"
            ],
            "example": "ACTIVE"
          },
          "createdAt": {
            "type": "string",
            "example": "2026-06-09T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "image",
          "role",
          "status",
          "createdAt"
        ]
      },
      "UpdateCurrentUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "sijun"
          }
        },
        "required": [
          "name"
        ]
      },
      "CommentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "comment-1"
          },
          "postId": {
            "type": "string",
            "example": "post-1"
          },
          "content": {
            "type": "string",
            "example": "좋은 기준입니다."
          },
          "authorId": {
            "type": "string",
            "example": "user-1"
          },
          "authorName": {
            "type": "string",
            "example": "sijun"
          },
          "createdAt": {
            "type": "string",
            "example": "2026-06-09T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-06-09T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "postId",
          "content",
          "authorId",
          "authorName",
          "createdAt",
          "updatedAt"
        ]
      },
      "CommentListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommentDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
      "CreateCommentDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "example": "좋은 기준입니다."
          }
        },
        "required": [
          "content"
        ]
      },
      "UpdateCommentDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "example": "수정된 댓글입니다."
          }
        }
      },
      "DeleteCommentResponseDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          },
          "id": {
            "type": "string",
            "example": "comment-1"
          }
        },
        "required": [
          "ok",
          "id"
        ]
      },
      "PostTagDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "tag-react"
          },
          "name": {
            "type": "string",
            "example": "react"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "PostDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "post-1"
          },
          "title": {
            "type": "string",
            "example": "보일러플레이트 구조 결정"
          },
          "excerpt": {
            "type": "string",
            "example": "라우터, 서버 상태, URL 상태를 분리한다."
          },
          "content": {
            "type": "string",
            "example": "게시판 기능은 스택 검증을 위한 최소 CRUD 화면으로 둔다."
          },
          "authorId": {
            "type": "string",
            "example": "user-1"
          },
          "authorName": {
            "type": "string",
            "example": "sijun"
          },
          "createdAt": {
            "type": "string",
            "example": "2026-06-09T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-06-09T00:00:00.000Z"
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostTagDto"
            }
          }
        },
        "required": [
          "id",
          "title",
          "excerpt",
          "content",
          "authorId",
          "authorName",
          "createdAt",
          "updatedAt",
          "tags"
        ]
      },
      "PostListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostDto"
            }
          },
          "page": {
            "type": "number",
            "minimum": 1,
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "minimum": 1,
            "example": 10
          },
          "totalItems": {
            "type": "number",
            "minimum": 0,
            "example": 3
          },
          "totalPages": {
            "type": "number",
            "minimum": 1,
            "example": 1
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "totalItems",
          "totalPages"
        ]
      },
      "CreatePostDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "프론트 공통 스택 기록"
          },
          "excerpt": {
            "type": "string",
            "example": "TanStack Router와 Query를 연결한다."
          },
          "content": {
            "type": "string",
            "example": "OpenAPI 기반 codegen으로 화면 개발의 출발점을 만든다."
          },
          "tagIds": {
            "default": [],
            "example": [
              "tag-react"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "title",
          "excerpt",
          "content"
        ]
      },
      "UpdatePostDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "수정된 제목"
          },
          "excerpt": {
            "type": "string",
            "example": "수정된 요약"
          },
          "content": {
            "type": "string",
            "example": "수정된 본문"
          },
          "tagIds": {
            "example": [
              "tag-react"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "DeletePostResponseDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          },
          "id": {
            "type": "string",
            "example": "post-1"
          }
        },
        "required": [
          "ok",
          "id"
        ]
      },
      "HealthResponseDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "ok"
        ]
      }
    }
  }
}